+++ src/libclipper/include/clipper/query_processor.hpp
@@ -56,16 +56,6 @@ class QueryProcessor {
       selection_policies_;
 };
 
-class PredictError : public std::runtime_error {
- public:
-  PredictError(const long query_id, const std::string msg);
-  const char* what() const noexcept;
-
- private:
-  const long query_id_;
-  const std::string msg_;
-};
-
 }  // namespace clipper
 
 #endif
